+++ src/yafowil/widget/ace/example.py
@@ -5,6 +5,8 @@ DOC_ACE_PYTHON = """
 ACE in Python mode
 ------------------
 
+Set the 'mode' option to 'python' to create a Python Ace widget.
+
 .. code-block:: python
 
     value = '''
@@ -44,6 +46,8 @@ DOC_ACE_JS = """
 ACE in Javascript mode
 ----------------------
 
+Set the 'mode' option to 'javascript' to create a JavaScript Ace widget.
+
 .. code-block:: python
 
     value = '''
@@ -81,5 +85,154 @@ def ace_js():
     }
 
 
+DOC_ACE_THEMES = """
+Themes
+------
+
+Customize syntax highlighting and appearance with one of the available ACE themes.
+
+.. code-block:: python
+
+    value = '''
+    function foo(items) {
+        for (var i=0; i<=items.length; i++) {
+            alert(items[i]);
+        }
+    }
+    '''
+    ace = factory('#field:ace', value=value, props={
+        'label': 'Solarized Theme',
+        'required': 'Code is required',
+        'theme': 'solarized_light',
+        'mode': 'javascript'
+    })
+"""
+
+ace_theme_value = """\
+function foo(items) {
+    for (var i=0; i<=items.length; i++) {
+        alert(items[i]);
+    }
+}
+"""
+
+
+def ace_themes():
+    part = factory(u'fieldset', name='yafowil.widget.ace.themes')
+    part['ace'] = factory('#field:ace', value=ace_theme_value, props={
+        'label': 'Solarized Theme',
+        'required': 'Code is required',
+        'theme': 'solarized_light',
+        'mode': 'javascript'
+    })
+    return {
+        'widget': part,
+        'doc': DOC_ACE_THEMES,
+        'title': 'Themes',
+    }
+
+
+DOC_ACE_DARKMODE = """
+Dark Mode
+---------
+
+Customize syntax highlighting and appearance for dark mode with one of the
+available ACE themes.
+The current ACE theme corresponds to the theme set in Bootstrap5 HTML
+data-bs-theme attribute ('light' | 'dark').
+
+Try it out using the color switch above!
+
+.. code-block:: python
+
+    value = '''
+    function foo(items) {
+        for (var i=0; i<=items.length; i++) {
+            alert(items[i]);
+        }
+    }
+    '''
+    ace = factory('#field:ace', value=value, props={
+        'label': 'Light and Dark Theme',
+        'required': 'Code is required',
+        'theme': 'chrome',
+        'dark_theme': 'tomorrow_night',
+        'mode': 'javascript'
+    })
+"""
+
+ace_darkmode_value = """\
+function foo(items) {
+    for (var i=0; i<=items.length; i++) {
+        alert(items[i]);
+    }
+}
+"""
+
+
+def ace_darkmode():
+    part = factory(u'fieldset', name='yafowil.widget.ace.darkmode')
+    part['ace'] = factory('#field:ace', value=ace_theme_value, props={
+        'label': 'Light and Dark Theme',
+        'required': 'Code is required',
+        'theme': 'chrome',
+        'dark_theme': 'tomorrow_night',
+        'mode': 'javascript'
+    })
+    return {
+        'widget': part,
+        'doc': DOC_ACE_DARKMODE,
+        'title': 'Dark Theme',
+    }
+
+
+DOC_ACE_DISPLAY = """
+Display Mode / Read-only Mode
+-----------------------------
+
+The widget attribute ``read_only`` can be used to render the widget in read-only mode.
+
+The widget's display renderer automatically sets the ``read_only`` attribute to ``True``.
+
+.. code-block:: python
+
+    value = '''
+    function foo(items) {
+        for (var i=0; i<=items.length; i++) {
+            alert(items[i]);
+        }
+    }
+    '''
+    ace = factory('#field:ace', value=value, mode='display', props={
+        'label': 'Light and Dark Theme',
+        'required': 'Code is required',
+        'theme': 'chrome',
+        'dark_theme': 'tomorrow_night',
+        'mode': 'javascript',
+        # 'read_only': True # manually render in read-only mode
+    })
+"""
+
+
+def ace_display():
+    part = factory(u'fieldset', name='yafowil.widget.ace.display')
+    part['ace'] = factory('#field:ace', value=ace_theme_value, mode='display', props={
+        'label': 'Display Mode',
+        'required': 'Code is required',
+        'theme': 'chrome',
+        'dark_theme': 'tomorrow_night',
+        'mode': 'javascript'
+    })
+    return {
+        'widget': part,
+        'doc': DOC_ACE_DISPLAY,
+        'title': 'Display Mode',
+    }
+
+
 def get_example():
-    return [ace_python(), ace_js()]
+    themes = []
+    if factory.theme == 'bootstrap5':
+        themes.append(ace_darkmode())
+    themes += [ace_python(), ace_js(), ace_themes(), ace_display()]
+    return themes