+++ CHANGES.rst
@@ -1,10 +1,49 @@
 Changes
 =======
 
-1.3 (unreleased)
-----------------
+2.0.0 (unreleased)
+------------------
 
-- No changes yet.
+- Introduce ``readonly`` widget attribute and support rendering in display mode.
+  [lenadax]
+
+- Update jQuery to version ``4.0.0-beta.2``.
+  [lenadax]
+
+- New widget configuration options: ``dark_theme``, ``wrapper_class``.
+  [lenadax]
+
+- Implement dark theme option for ``Bootstrap5`` dark theme mode.
+  [lenadax]
+
+- Use rollup for bundling scss. Use ``make rollup`` to compile js and scss.
+  [lenadax]
+
+- Use ``webtestrunner`` instead of ``karma`` for js tests. Use ``make wtr`` to run tests.
+  [lenadax]
+
+- Use ``pnpm`` as package manager.
+  [lenadax]
+
+- Create Bootstrap5 widget version.
+  [lenadax]
+
+
+2.0a1 (2023-05-15)
+------------------
+
+- Add ``webresource`` support.
+  [rnix]
+
+- Extend JS by ``ace_on_array_add`` and ``register_array_subscribers``
+  to enable usage in ``yafowil.widget.array``.
+  [lenadax]
+
+- Update ACE Editor to v1.4.13.
+  [rnix]
+
+- Rewrite JavaScript using ES6.
+  [rnix]
 
 
 1.2 (2018-07-16)